+++ src/main/java/fr/alasdiablo/mods/factory/recycling/data/tag/RecyclingFactoryBlockTagsProvider.java
@@ -1,0 +1,30 @@
+package fr.alasdiablo.mods.factory.recycling.data.tag;
+
+import fr.alasdiablo.mods.factory.recycling.RecyclingFactory;
+import fr.alasdiablo.mods.factory.recycling.init.RecyclingFactoryBlocks;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
+import net.minecraft.tags.BlockTags;
+import net.neoforged.neoforge.common.data.BlockTagsProvider;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.concurrent.CompletableFuture;
+
+public class RecyclingFactoryBlockTagsProvider extends BlockTagsProvider {
+    public RecyclingFactoryBlockTagsProvider(
+            PackOutput output,
+            CompletableFuture<HolderLookup.Provider> lookupProvider,
+            @Nullable ExistingFileHelper existingFileHelper
+    ) {
+        super(output, lookupProvider, RecyclingFactory.MODID, existingFileHelper);
+    }
+
+    @Override
+    protected void addTags(HolderLookup.@NotNull Provider provider) {
+        this.tag(BlockTags.MINEABLE_WITH_PICKAXE)
+                .add(RecyclingFactoryBlocks.RUBBISH_BIN.get())
+                .add(RecyclingFactoryBlocks.STIRLING_RECYCLING_CRUSHER.get());
+    }
+}