+++ src/main/java/fr/alasdiablo/mods/factory/recycling/data/tag/RecyclingFactoryItemTagsProvider.java
@@ -1,0 +1,28 @@
+package fr.alasdiablo.mods.factory.recycling.data.tag;
+
+import fr.alasdiablo.mods.factory.recycling.RecyclingFactory;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
+import net.minecraft.data.tags.ItemTagsProvider;
+import net.minecraft.data.tags.TagsProvider;
+import net.minecraft.world.level.block.Block;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.concurrent.CompletableFuture;
+
+
+public class RecyclingFactoryItemTagsProvider extends ItemTagsProvider {
+    public RecyclingFactoryItemTagsProvider(
+            PackOutput output,
+            CompletableFuture<HolderLookup.Provider> lookup,
+            @NotNull TagsProvider<Block> blockTagsProvider,
+            @Nullable ExistingFileHelper existingFileHelper
+    ) {
+        super(output, lookup, blockTagsProvider.contentsGetter(), RecyclingFactory.MODID, existingFileHelper);
+    }
+
+    @Override
+    protected void addTags(HolderLookup.@NotNull Provider provider) {}
+}