+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### 2.0.0
 - Refactor the source code to use UK English
+- Add Stirling Recycling Crusher
+- Fix missing loot table
+- Fix missing tags
 
 ### 1.2.0
 - Add a config system to allow adding and removing entry in the scrap box loot table